+++ scripts/macros/monsterFeatures/shamblingMound/shamblingMound.js
@@ -1,0 +1,6 @@
+import {engulf} from './engulf.js';
+import {slam} from './slam.js';
+export let shamblingMound = {
+    'engulf': engulf,
+    'slam': slam
+}